crm-5: refactoring permissions to his trait
