+++ app/Models/User.php
@@ -2,21 +2,18 @@
 
 namespace App\Models;
 
-// use Illuminate\Contracts\Auth\MustVerifyEmail;
-
+use App\Models\Traits\HasPermissions;
 use Illuminate\Auth\Passwords\CanResetPassword;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
-use Illuminate\Support\Facades\Cache;
 
 class User extends Authenticatable
 {
     use HasFactory;
     use Notifiable;
     use CanResetPassword;
+    use HasPermissions;
 
     protected $fillable = [
         'name',
@@ -35,34 +32,5 @@ class User extends Authenticatable
             'email_verified_at' => 'datetime',
             'password'          => 'hashed',
         ];
-    }
-
-    public function permissions(): BelongsToMany
-    {
-        return $this->belongsToMany(Permission::class);
-    }
-
-    public function givePermissionTo(string $key): void
-    {
-        $this->permissions()->firstOrCreate(compact('key'));
-
-        $cacheKey = $this->getPermissionsCacheKey();
-        Cache::forget($cacheKey);
-        Cache::rememberForever($cacheKey, fn () => $this->permissions);
-    }
-
-    public function hasPermissionTo(string $key): bool
-    {
-        /** @var Collection $permissions */
-        $permissions = Cache::get($this->getPermissionsCacheKey(), $this->permissions);
-
-        return  $permissions
-            ->where('key', '=', $key)
-            ->isNotEmpty();
-    }
-
-    private function getPermissionsCacheKey(): string
-    {
-        return 'user::' . $this->id . '::permissions';
     }
 }
